math: forbid dubious inequality comparisons.
The issue, reported by Paul A. Patience, is that
code like (< 1 "abc") is successfully producing a result.
The root cause is that 1 is an iterable object, and so
is treated as a sequence opposite to the "abc" operand.

We should allow only true sequences in this situation.

* arith.c (seq_lt_compat_check): New static function. Checks
that neither of two sequences is SEQ_NOTSEQ or SEQ_HASHLIKE.
(seq_lt, seq_le): Use seq_lt_compat_check to reject dubious
inputs.

* txr.1: Minor wording change in the related documentation,
removing a gratuitous adjective.

* tests/016/arith.tl: Inequality tests.