system: stop using popen if we have fork.
On all the platforms on which TXR is regularly built, there is
fork, including the Windows port via Cygwin. We therefore
don't have to be using popen for the open-command function.
The handling the stream redirection will be more efficient if
we implement open-command in terms of open-process, and the
special mode string extensions for redirection like ">21"
will work, too.

* stream.c (fds_subst, fds_swizzle): Define these functions
only for !HAVE_FORK_STUFF platforms.
(open_command): popen-based implementation moved down into a
section of code for !HAVE_FORK_STUFF platforms. A separate
implementation is written for HAVE_FORK_STUFF platforms
which passes the command to an appropriate Windows native
or POSIX interpreter.

* utf8.c (w_popen): Nothing but open-command calls this
function, so we don't need it on HAVE_FORK_STUFF platforms.
Wrap with #if.

* utf8.h (w_popen): wrap with #if.