suffix functions: requirements change.
The short-suffix and long-suffix functions will now return the
suffix including the leading period. This was a suggestion
from user Paul A. Patience, which is a good requirement.

Since these functions were newly introduced just the last
release, I'm not going to provide backwards compatibility
switching for them.

* stream.c (short_suffix, long_suffix): Duplicate the suffix
starting at the dot, not dot + 1.

* tests/018/path.tl: Test cases updated.

* txr.1: Documentation updated.