linenoise: end-of-line/buffer selection glitch.
In inclusive selection mode, when the selection is reversed
(end point is before start), and the starting character
is the end of the line or of the buffer, that character
is not included in the highlight, as if non-inclusive
selection were in effect.

This doesn't affect the semantics of the selection, only
the way it is rendered visually; the character which is
not highlighted is still included in the selection.

* linenoise/linenoise.c (sync_data_to_buf): Remove two bogus
conditions from the line which extends the visual selection
by one character: we must not avoid executing this logic if
the current character is zero (end of buffer) or CR (end of line).