regex: bugfix: print/read consistency of n-ary ops.
The regex-compile function accepts syntax in which certain
operators like (or ...) can be n-ary. This representation
is converted to the strictly binary form that is understood
by the compiler internals (and which regex-parse outputs),
as well as the regex printer.

Unfrotunately, regex-compile is attaching the original
form with the n-ary operators to the regex object, and the
printer does not reat this; it renders the syntax with
portions missing. It needs the binary form.

* regex.c (regex_compile): Capture the intermediate result
from calling reg_nary_to_bin, and use that as regex_source.
The pass that through the remaining two optimization passes
to obtain regex_sexp which is compiled.