main: maintain -b option.
* txr.c (help): Remove -b help text which pertains to an
ancient -b option that was deprecated in 2014. Add correct
help text for -b.
(txr_main): Use split_str for the argument of -b. If it fails
to produce a two-element list, it means the equal sign is
missing; issue an error message and exit. Thereby we avoid
passing nil as the input source argument of lisp_parse.