signal: fix typo causing missing itimer-prof.
* signal.c (sig_init): Register ITIMER_PROF as the documented
itimer-prof, not the undocumented itimer-prov. Anyone
depending on this will have to fix their code to use the new
variable, or else (defvarl itimer-prof itimer-prov).
This was discovered due to Paul A. Patience's doc-syms
scrubbing code recently added to genman.txr, which identifies
documented symbols not present in the actual image.