main: revise error-ignore strategy for Lisp file.
txr can process a Lisp file specified on the command line and
then enter into the listener. In order to always enter the
listener, even if the file errors out, the execution of the
file uses read_eval_stream_noerr, which was written for this
purpose. But this causes a problem for error reporting which
is sensitive to whether or not an error handler exists.
(See the compile-error function, and how it uses find-frame.)

* txr.c (txr_main): Because we already know whether we are
going to be entering the listener, we can split the evaluation
of the Lisp file into two cases: if we will be entering the
listener, we evaluate it with ignored error exceptions.
If we won't be entering the listener, we evaluate normally.
With this, I can run "make tests" now and be taken to the
locaton of certain errors, due to the messages appearing on
standard error.