regex-compile: argument defaulting issue.
* regex.c (regex_compile): In the string case when regex_parse
needs to be called, redundant argument defaulting is being
performed on error_stream, defaulting it to nil. The
regex_parse function doesn't like that; it no longer accepts
nil as a facsimile of missing.  The right thing to do here is
nothing: just pass error_stream to regex_parse as-is and let
it do the defaulting. And that is how the code was initially
until commit a8b0e36b1760e51a8a3a25d4e22a325e407ef3d4 in March
of 2014, which introduced the redundant defaulting.
Reported by Paul A. Patience