lib: default_arg_strict becomes macro.
* lib.h (default_arg_strict): Inline function converted to
macro, so that we can suppress the evaluation of the default
expression if it is not required. The default_arg_strict idiom
is being used all over the stream library now. The common
default value expressions like std_out are macros that access
the dynamic variable by symbol. We don't want to be wastefully
doing that on each stream operation, whether or not a stream
argument has been supplied.