op: fix bug in do.
The June 30 09e70c914ca83b5c7405aa633864db49f27efa05,
subject "op: refactor do handling", introduced a regression
breaking the tags.tl program. An implicit argument
gets inserted twice:

   [[(do op list @1)] 'x] -> (x x) ;; incorrect/weird

This was spotted by Paul A. Patience while working on
extending tags.tl for Emacs.

It's not exactly a regression because the original
behavior is not documented or tested, and has issues;
we simply cannot roll back the commit; a proper
fix is required.

How the above call is now supposed to work is that:

- the @1 parameter belongs to the op, not to the do.

- the do therefore has no explicitly given parameters
  of its own.

- therefore the do inserts its parameter.

In other words (do op list @1) is formally equivalent
to (do op list @1 @@1).

Both levels of function indirection require an
argument:

  [[(do op list @1) 'x] 'y] -> (y x)

  [[(do op list @1 @@1) 'x] 'y] -> (y x)

* stdlib/op.tl (sys:op-ctx): The structure gets a new slot,
nested, which is a flag indicating whether unprocessed nested
metas occur. This is critically needed because the
sys:op-alpha-rename passes which are called with
do-nested-metas being false do not insert nested metas into
the gens list; they transform them and leave them in the
syntax. Yet we must make decisions based on their presence.
Conretely, we must be able to tell that (do op list @@1)
has a meta against the outer (do ...), while we are just
processing the do.
(sys:op-alpha-rename): When replacing a nested meta syntax
with the macro invocation, we set the nested flag of the
parent context true.
(sys:op-expand): Bring back the do-gen; we need it. We cannot
simply insert @1 into the syntax, because that is not
lexically transparent. If we add @1 to (do op ...) then that
@1 is interpreted as belonging to the op, not to the do.
We must also check the new Boolean flag nested to properly
detect whether we have metas, including unexpanded nested
metas.

* tests/012/op.tl: New test cases combining (do op ...).