op: consolidate testing for @rec/@(rec ...)
* stdlib/op.tl (sys:op-rec-p): Test for both the variable @rec
and the expresson @(rec ...). Also, the expr argument of the function is
renamed to exp because it causes a problem, due to that symbol,
sys:expr, being the meta indicator.
(sys:op-ensure-rec): Remove one test for (sys:var usr:rec), since
sys:op-rec-p does it. Convert one equal-test to a call to sys:op-rec-p.