tags: don't escape etag patterns
The first field of each etag definition is referred to in the spec as
the "pattern", but it is supposed to contain literal text, and therefore
no characters within it need be escaped.

* tags.tl (escape): Move above tag definition.
(tag)[pattern]: Rename to...
[line]: ...this.
[text]: Update renamed slot. Escape the line here rather than on
creation.
[etext]: Update renamed slot.
(slot-tag)[text]: Update renamed slot. Escape the line here rather than
on creation.
(orig-tag)[line]: Rename to...
[orig-line]: ...this.
[text]: Update renamed slot.
(get-pos-pat): Rename to...
(get-pos-line): ...this. Don't escape the line when returning it.
(with-tag-shorthand-macro, toplevel): Rename variables and references to
functions in accordance with the above.