gc: problem in environment-copying functions.
* eval.c (copy_env, deep_copy_env): These functions are not following
a protocol for object construction that is correct under generational
GC. They are allocating a new object with make_obj first, and then
calling functions to copy the constituent elements to populate into
the object with a direct assignment. This direct assignment is wrong;
the set macro is required. A better fix, rather than using the set
macro, is to copy the constituent parts first, holding them in local
variables, then allocate the new object, and finally, without doing any
other memory allocating operations, assign the constituent parts into
the new object.