streams: bad argument defaulting in close-stream.
* stream.c (stdio_close, pipe_close): Fix throw_on_error
argument not being defaulted correctly, so that errors are
thrown even when the argument is omitted.

* strudel.c (strudel_close): Here, we also must default the
argument. The corresponding close method does not have an
optional argument; it is mandatory. The documentation is
bungled for it, though.

* txr.1: Fix documentation of structure delegate streams with
regard to the close method. It does not take offs and whence
parametrs, but throw-on-error-p, which is mandatory.