listener: prompt feature for plain mode.
The :prompt-on command will enable prompting in plain mode.

* linenoise/linenoise.c (struct lino_state): New member, show_prompt.
(line_enable_noninteractive_prompt): New function.
(linenoise): In the plain mode loop, the show_prompt flag is on, show
the prompt. For continuation lines, show a condensed prompt, which
consists of the suffix of the full prompt, starting on the last
non-whitespace character.

* linenoise/linenoise.h (lino_enable_noninteractive): Declared.

* parser.c (repl): Implement :prompt-on command which enables the
above mode.

* txr.1: Documented.