parser: @(mdo) must not be subject to expand-meta.
* parser.y (elem): When the elem is a list, if it starts with mdo, then
we evaluate it immediately and substitute (do) as the semantic
value.  We no longer not allow mdo to precipitate into
match_expand_elem, where expand_meta will be unleashed on it.
Doing so causes the bug that expessions like @1, denoting
the form (sys:var 1), are rewritten to (sys:expr 1), which op
syntax. So two bugs are fixed: the incorrect treatment of
meta-syntax, and the neglect to perform mdo in nested
contexts.
(check_parse_time_action): We need not handle mdo here any
more; it will never make it into this function.
Only actions done in the main clause list belong here,
not parse time actions done at any nesting level.

* tests/008/mdo.txr: New file.

* y.tab.c.shipped: Updated.