format: bug: sign not reset before each conversion.
For instance, this bad output is produced:

  (pic "+0####.## <<<<<" 123 1) -> "+00123.00 +1   "

The second argument should not have any leading + sign.

* stream.c (formatv): For each new conversion specifier introduced by ~,
reset the sign variable to zero also. That's the semantic change here,
occluded by the fact that I'm rearranging the declaration of the
variables, adding comments, and condensing the assignments while also
getting them into the same order as the declarations, in order that this
sort of bug does not creep in in the future, should another such
variable be added.

* tests/018/format.tl: Adding correct version of above test case.