sub: allow generic iterables.
In this patch we allow the s in (sub s from to) and [s from..to]
to be any iterable.

* lib.c (iter_dynamic, sub_iter): New static function.
(generic_funcall): Handle all objects via the sequence case:
ref, sub and all that. Unfortunately, we lose some error
handling at the level of the sub function. But we allow any
iterable to be passed through to sub.
(sub): Handle default case through sub_iter.

* tests/012/iter.tl: New cases.

* txr.1: Documented.