ffi: ffi macro uses load-time.
A number of functions take an argument which is a ffi type.
Typically, this argument is produced using by a
ffi-type-compile call which is produced by the ffi macro. But
this ffi-type-compile call is invoked at run time, each time
such a function is called. A solution for this is to have the
ffi macro hoist the compilation to load time.

* stdlib/ffi.tl (ffi): Add load-time wrapping to generated
expression.

* txr.1: Updated correspondence between (ffi ...) form and
equivalent (ffi-type-compile form).